[querier] move profile dictionary
- profile as querier application
